update rest and componenetes

+++ src/POC/ui-components/PropertyCreateForm.jsx
@@ -362,9 +362,9 @@ export default function PropertyCreateForm(props) {
         step="any"
         value={longitud}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               streetName,
@@ -398,9 +398,9 @@ export default function PropertyCreateForm(props) {
         step="any"
         value={latitude}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               streetName,
