+++ src/POC/ui-components/PropertyUpdateForm.jsx
@@ -378,9 +378,9 @@ export default function PropertyUpdateForm(props) {
         step="any"
         value={longitud}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               streetName,
@@ -414,9 +414,9 @@ export default function PropertyUpdateForm(props) {
         step="any"
         value={latitude}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               streetName,
